Add build documentation (#55)
This PR adds documentation of the RAPIDS build infrastructure to this
repository. This is just a first pass, but since these are purely
internal-facing I think it's easiest for us to start with something and
iterate. These are meant for the team to document and evolve our
practices over time. The docs may be built using
[mkdocs](https://www.mkdocs.org/), which I chose because it is very
lightweight and uses pure Markdown (I don't think the overhead of Sphinx
is merited here).